Second push hoping unsigned int solves everything

+++ C03/ex01/ft_strncmp.c
@@ -6,18 +6,18 @@
 /*   By: sawijnbe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/14 13:44:00 by sawijnbe          #+#    #+#             */
-/*   Updated: 2025/07/14 19:41:03 by sawijnbe         ###   ########.fr       */
+/*   Updated: 2025/07/14 21:31:31 by sawijnbe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_strncmp(char *s1, char *s2, unsigned int n)
 {
-	int	i;
+	unsigned int	i;
 
 	if (s1 == s2)
 		return (0);
 	i = -1;
-	while ((s1[++i] || s2[i]) && i < (int)n)
+	while ((s1[++i] || s2[i]) && i < n)
 		if (s1[i] != s2[i])
 			return (s1[i] - s2[i]);
 	return (0);
